requirements.txt
easydict==1.11
loguru==0.7.2
matplotlib==3.8.2
Pillow==9.5.0
PyYAML==6.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
onnx==1.15.0
onnx-tool==0.9.0
